fix pagestate styles
